+++ utilities/poc_local_storage/lib/protected_screen.dart
@@ -1,0 +1,18 @@
+import 'package:flutter/material.dart';
+
+class ProtectedScreen extends StatelessWidget {
+  const ProtectedScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('Protected Screen'),
+        backgroundColor: Theme.of(context).colorScheme.inversePrimary,
+      ),
+      body: const Center(
+        child: Text('Welcome to the protected screen!'),
+      ),
+    );
+  }
+}
