+++ utilities/poc_local_storage/pubspec.yaml
@@ -5,18 +5,19 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=3.4.3 <4.0.0"
+  sdk: ">=3.5.0 <4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
+  go_router: ^14.2.3
   file_picker: ^8.0.7
   flutter_secure_storage: ^9.2.2
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^3.0.0
+  flutter_lints: ^4.0.0
 
 dependency_overrides:
   flutter_secure_storage_web: ^2.0.0-beta.1
